MFLT-5638 cmake recipe for expat version conflict
The conda-forge cmake package has cmake requirement of:

`expat >=2.4.1,<3.0a0`

This conflicts with our internally built gdb packages, which are
(incorrectly) set to ^2.2:

`expat >=2.2.10,<2.3.0a0`

It's simpler to just fork cmake especially since we only use it in
testing, not in production.

Recipe taken from here:

https://github.com/conda-forge/cmake-feedstock/tree/284e91676110b5d52130bd6fb04e91f099194fcf/recipe

Only modification is to specify `expat: '2.2'` instead of `expat: '2'`.