Add github action for opt-in building linux+osx
It's a bummer trying to cross build for osx on linux 😣. Add a github
action that uses the github-hosted macos runner for building osx conda
packages.

It's opt-in, have to set the package dir to enable it in a pull request.

May not work well with the places we've set `CONDA_BUILD_SYSROOT` (noted
in the readme).