Add patch for fixing GDB's disassemble for 10+ (#20)
### Get Conda to build me GDB with symbols

I didn't really want to create an environment that could properly build GDB for my Mac machine, so I let Conda do it.
I inserted an `exit 1` in the `build.sh` script so that it failed at the exact time. When a Conda build fails, it leaves the work directly alone so that a developer can `cd` into it and debug it.

Upon building the GDB, the gdb binary will be located at `....../work/gdb/gdb`

### Debugging

So you can debug it with our local GDB and tell it to load the symbol file which breaks:
```
$ cd $HOME/mambaforge/envs/build/conda-bld/multi-arch-gdb_1658520479521/work/gdb/
$ /usr/local/bin/gdb --args ./gdb $HOME/dev/memfault/memfault/tools/tests/fixtures/binaries/symbols/heap-analyzer-example.elf

# Now run it
(gdb) run

# ... nested GDB now

(gdb) disassemble 0x8003c44,+2
```

The line it was segfaulting on was https://github.com/bminor/binutils-gdb/blob/master/opcodes/arm-dis.c#L11867-L11868.
Turns out that `info->symtab` was NULL, and it was dereferencing that pointer.

### WHY!?

In a change from 3 years ago, apparently there was a check for this `symtab` being NULL (rather checking `symtab_size == 0`, but same thing), but it was removed in this commit.

https://github.com/bminor/binutils-gdb/commit/796d6298bb11deab06814cc38cfe74a1bfc57551#diff-d2da7fc6562c73a3b97a5f9eaf58d5aad41e997853c8ace07c532a04365f61ccR6371

Adding it back, everything seems to work again for us. So my short term solution is just to create a patch with this back in place and say ship it.