Build esp32 gdb multibin from source (#28)
Update the recipe to build with the `--with-xtensaconfig` custom
configuration and library that enables debugging esp32+esp32s2+esp32s3
with a single gdb binary. This was made possible thanks to Espressif
open sourcing the necessary `xtensaconfig` submodule repo:

https://github.com/espressif/esp-idf/issues/9459

Version updated to `memfault1` from `memfault0` to distinguish this
flavor from the previous version.

I've tested on linux-x86_64 and mac-x86_64.