gdb multi-arch v13.1
Since it's now released! Can also drop our old patch for the segfault
issue on certain ARM backtraces 🙌

I took the opportunity to rename the output binaries to have the
`multi-arch-` prefix, because it's really confusing to me to have `gdb`
not be native `gdb` when in a conda environment with this package
installed.

I've also added python3.11 to the build matrix, now that it's released.

And include `readelf` in the package (`multi-arch-readelf`), since it
can be pretty handy.

Finally, `shfmt` the build script too since I'm in here.