fix(volta): allow activate.sh to work in strict mode
  ### Summary

If the script is sourced when run in strict mode (`-u`) (disallowing unbound variables, the activate.sh script would fail because `VOLTA_HOME` was unbound. This PR adds a fallback to avoid this error.

  ## Test plan

Add `set -u` at the top of the script and run it. Observe it no longer fails.