gdb multi-arch v14.1
Since it's now released! Can also drop our old patch for the segfault
issue on certain ARM backtraces 🙌

I took the opportunity to rename the output binaries to have the
`multi-arch-` prefix, because it's really confusing to me to have `gdb`
not be native `gdb` when in a conda environment with this package
installed.

Refreshed the `conda_build_config.yaml` to use the latest conda-build
feedstock version at time of writing:

```bash
❯ curl -sSL https://github.com/conda-forge/conda-forge-pinning-feedstock/raw/main/recipe/conda_build_config.yaml > gdb-multi-arch/conda_build_config.yaml
```

Which is
https://github.com/conda-forge/conda-forge-pinning-feedstock/blob/41ef3139dd637914a18f9609bf3d33a2f4780376/recipe/conda_build_config.yaml
.

This changes the supported + built python targets to:

- 3.8
- 3.9
- 3.10
- 3.11

And include `readelf` in the package (`multi-arch-readelf`), since it
can be pretty handy.

Finally, `shfmt` the build script too since I'm in here.