chore(aws-vault): Update aws-vault to 7.2.0 (#70)
### Issues addressed

aws-vault added support for `sso-session` in 7.0.0, and I'd like to
use this feature to reduce boilerplate.

I've also noticed a few rough edges which may have been ironed out in
the past 2 years since our last version bump.

 ### Summary

Version bump!

I built for linux and uploaded:

```
$ anaconda --user <redacted> --token <redacted> upload /home/topher/mambaforge/envs/build/conda-bld/linux-64/aws-vault-7.2.0-1.conda
```

I'm not sure how to build for the two Darwin builds... can someone give
me a hint? (or does someone with a mac have to do it? that feels like it
might be likely).

 ### Test Plan

- Will run locally for a bit to ensure it works as expected.